fix: move utils into internal folder
