+++ internal/utils/utils.go
@@ -1,4 +1,4 @@
-package internal
+package utils
 
 import (
 	"sort"
